pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["EvgenArt"]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.8.1"
PyYAML = "^6.0"
pytest = "^7.2.0"
flake8 = "^5.0.4"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
